+++ ImageEvaluator/ImageEvaluatorLib/CalculateStatisticalData/CalculateColumnDataBase.cs
@@ -53,7 +53,7 @@ namespace ImageEvaluatorLib.CalculateStatisticalData
         /// <param name="meanVector"></param>
         /// <param name="stdVector"></param>
         /// <returns></returns>
-        public abstract bool Run(Image<Gray, float> inputImage, Image<Gray, byte> maskImage, int[,] pointArray, ref Image<Gray, float> meanVector, ref Image<Gray, float> stdVector);
+        public abstract bool Run(Image<Gray, ushort> inputImage, Image<Gray, byte> maskImage, int[,] pointArray, ref Image<Gray, float> meanVector, ref Image<Gray, float> stdVector);
 
 
         /// <summary>
@@ -65,7 +65,7 @@ namespace ImageEvaluatorLib.CalculateStatisticalData
         /// <param name="meanVector"></param>
         /// <param name="stdVector"></param>
         /// <returns></returns>
-        protected virtual bool CheckInputData(Image<Gray, float> inputImage, Image<Gray, byte> maskImage, int[,] pointArray, Image<Gray, float> meanVector, Image<Gray, float> stdVector)
+        protected virtual bool CheckInputData(Image<Gray, ushort> inputImage, Image<Gray, byte> maskImage, int[,] pointArray, Image<Gray, float> meanVector, Image<Gray, float> stdVector)
         {
             try
             {
